Styling header & project sections
