+++ src/js/index.js
@@ -6,6 +6,7 @@ import { projectsList } from '../components/projects';
 import { TextScramble } from '../components/TextPhrases';
 import { Jokes } from '../components/ChuckJokes';
 import { UI } from '../components/ChuckUi';
+import '../components/DarkTheme';
 import Swal from 'sweetalert2';
 
 import { validateName, validateEmail } from '../components/isValid';
@@ -24,10 +25,7 @@ const main = new MainContainerComponent('main');
 // Projects
 projectsList(data);
 
-// ——————————————————————————————————————————————————
 // Text phrases
-// ——————————————————————————————————————————————————
-
 const phrases = [
   'The only way to do great work',
   'is to love what you do',
@@ -83,39 +81,6 @@ function getJokes(e) {
 const year = new Date();
 const date = document.querySelector('.date');
 date.textContent = year.getFullYear();
-
-// dark theme
-const themeSwitch = document.getElementById('switch');
-if (themeSwitch) {
-  initTheme(); // on page load, if user has already selected a specific theme -> apply it
-
-  themeSwitch.addEventListener('change', function (e) {
-    resetTheme(); // update color theme
-  });
-
-  function initTheme() {
-    const darkThemeSelected =
-      localStorage.getItem('themeSwitch') !== null &&
-      localStorage.getItem('themeSwitch') === 'dark';
-    // update checkbox
-    themeSwitch.checked = darkThemeSelected;
-    // update body data-theme attribute
-    darkThemeSelected
-      ? document.body.classList.add('dark')
-      : document.body.classList.remove('dark');
-  }
-
-  function resetTheme() {
-    if (themeSwitch.checked) {
-      // dark theme has been selected
-      document.body.classList.add('dark');
-      localStorage.setItem('themeSwitch', 'dark'); // save theme selection
-    } else {
-      document.body.classList.remove('dark');
-      localStorage.removeItem('themeSwitch'); // reset theme selection
-    }
-  }
-}
 
 // nav shadow
 window.addEventListener('scroll', (e) => {
