+++ src/scss/vars.scss
@@ -2,12 +2,16 @@
 
 // font
 $primary-font: 'Lato';
-$secondary-font: monospace, Helvetica, Arial, sans-serif;
+$secondary-font: monospace,
+Helvetica,
+Arial,
+sans-serif;
 
 // colors
-$primary-color: #202020;
-$secondary-color: #fcfcfc;
-$highlight-color: #80C6E4;
+$primary-color: #141e30;
+$lighten-primary-color: #243b55;
+$secondary-color: #fff;
+$highlight-color: #FA7564;
 
 $light-purple-color: #AC80E4;
 $light-red-color: #E48080;
